# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Tracker API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/comments/{taskId}": {
      "get": {
        "summary": "List Comments",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create Comment",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
  }
}